Improved portability for 32-bits systems
Ignore-this: cd3f4a5a182ab80d25b732938940e2c3

darcs-hash:c5b9680449ca8ed4f3a3f8ae9307176523e11d99

+++ src/common.h
@@ -10,7 +10,7 @@
 
 @define pi 3.14159265358979
 @undef HUGE
-@define HUGE 1e30
+@define HUGE ((double)1e30)
 @define nodata DBL_MAX
 
 @define max(a,b) ((a) > (b) ? (a) : (b))
